Proposed hack to filter out FM bearers with ISO country codes from SI.
FM bearers such as `fm:gb.c123.09580` where included and therefore considered valid in the legacy XSI specification, but are now no longer acceptable as part of the deprecation of using an ISO code in place of working out a GCC value based on geographic location.

XSLT v1 is unable to use RexEx to match a value, so this is a simplified alternative, although far from ideal.

Presented mostly for discussion.